+++ apps/swirl-docs/CHANGELOG.md
@@ -1,5 +1,34 @@
 # swirl-docs
 
+## 0.1.131
+
+### Patch Changes
+
+- Updated dependencies
+  [[`0823fc5`](https://github.com/getflip/swirl/commit/0823fc502c94345701350d4bdc9960ea7fccd8cc)]:
+  - @getflip/swirl-tokens@2.3.0
+  - @getflip/swirl-components@0.69.2
+  - @getflip/swirl-components-react@0.68.2
+
+## 0.1.130
+
+### Patch Changes
+
+- Updated dependencies
+  [[`7e525f0`](https://github.com/getflip/swirl/commit/7e525f0aa1ff210d0ac0ea771679ca6c01867930)]:
+  - @getflip/swirl-components@0.69.1
+  - @getflip/swirl-components-react@0.68.1
+
+## 0.1.129
+
+### Patch Changes
+
+- Updated dependencies
+  [[`d14251d`](https://github.com/getflip/swirl/commit/d14251dc1c599000333f99c135e67c9c9d729ad4)]:
+  - @getflip/swirl-components@0.69.0
+  - @getflip/swirl-icons@0.18.0
+  - @getflip/swirl-components-react@0.68.0
+
 ## 0.1.128
 
 ### Patch Changes